split objects with ts ,add interface
