+++ src/pages/repositories/repository/repo-comp/objects/useInterval.tsx
@@ -1,0 +1,19 @@
+import{useEffect, useRef} from "react";
+
+export const useInterval = (callback:()=>void, delay:number) => {
+    const savedCallback: React.MutableRefObject<()=>void> = useRef(()=>{});
+
+    useEffect(() => {
+        savedCallback.current = callback;
+    }, [callback]);
+
+    useEffect(() => {
+        function tick() {
+            savedCallback.current();
+        }
+        if (delay !== null) {
+            const id = setInterval(tick, delay);
+            return () => clearInterval(id);
+        }
+    }, [delay]);
+}
