+++ src/pages/repositories/services/import_data.tsx
@@ -10,6 +10,7 @@ const ImportPhase = {
     InProgress: 1,
     Completed: 2,
     Failed: 3,
+    Merging: 4
 }
 
 const startImport: IStartImport = async (setImportID, prependPath, commitMsg, sourceRef, repoId, refId, metadata = {}) => {
@@ -95,7 +96,13 @@ const ImportForm: React.FC<ImportFormProps> = ({
                     }) => {
     const [isSourceValid, setIsSourceValid] = useState(true);
     const importValidityRegexStr = config.import_validity_regex;
-    const storageNamespaceValidityRegex = RegExp(importValidityRegexStr);
+    let storageNamespaceValidityRegex:RegExp
+    if(importValidityRegexStr){
+        storageNamespaceValidityRegex = RegExp(importValidityRegexStr);
+    }else{
+        throw new Error("err ! type of importValidityRegexStr")
+    }
+    
     const updateSourceURLValidity = () => {
         if (sourceRef.current) {
             const isValid = storageNamespaceValidityRegex.test(sourceRef.current.value);
@@ -108,7 +115,7 @@ const ImportForm: React.FC<ImportFormProps> = ({
         }
         
     };
-    const sourceURIExample = config ? config.blockstore_namespace_example : "s3://my-bucket/path/";
+    const sourceURIExample= config ? config.blockstore_namespace_example : "s3://my-bucket/path/";
     return (<>
         <Alert variant="info">
             Import doesn&apos;t copy objects. It only creates links to the objects in the JiaoziFS metadata layer.
@@ -136,14 +143,14 @@ const ImportForm: React.FC<ImportFormProps> = ({
                 <Form.Group className='form-group'>
                     <Form.Label><strong>Destination:</strong></Form.Label>
                         <Form.Control type="text" autoFocus name="destination" ref={destRef} defaultValue={path}/>
-                    <Form.Text style={{color: 'grey'}} md={{offset: 2, span: 10000}}>
+                    <Form.Text style={{color: 'grey'}}>
                         Leave empty to import to the repository&apos;s root.
                     </Form.Text>
                 </Form.Group>
             }
             <Form.Group className='form-group'>
                 <Form.Label><strong>Commit Message:</strong></Form.Label>
-                <Form.Control sm={8} type="text" ref={commitMsgRef} name="commit-message" autoFocus defaultValue={`Imported data from ${config.blockstore_type}`}/>
+                <Form.Control  type="text" ref={commitMsgRef} name="commit-message" autoFocus defaultValue={`Imported data from ${config.blockstore_type}`}/>
             </Form.Group>
             <MetadataFields metadataFields={metadataFields} setMetadataFields={setMetadataFields}/>
             {err &&
